requirements.txt
langchain==0.0.219
openai==0.27.8
Pyrogram==2.0.106
TgCrypto==1.2.5
Psycopg2==2.9.6
python-dotenv==1.0.0